Add HostingEnivronment
We need to be able to configure the application in different
environments - mainly dev and 'production' ATM.

Environments are modelled in their own class - HostingEnivronment. This
stops config spreading across the app and gives us a simple way to set
policies about what happens in dev/staging and production.

The main draw is to add a sandbox mode to allow running the project
without external dependencies - Notify and the form API.

This commit stops notify being run if SANDBOX=true env var is set before
running the project.

Inspired by
https://github.com/DFE-Digital/apply-for-teacher-training/blob/main/app/lib/hosting_environment.rb