Add validation when retreiving answer
When a previous answer is restored from the session, invalid data can
cause issues. This happens if the question has changed or the validation
has been updated. This *should* only happen in development, but ATM it's
also possible in production if the form-creator changes the question
type.

Now, when we fetch an answer from the session, validate it again to make sure
it's still valid. If it's not, we use a blank question as though nothing
had been input before.

The code for this is messy and not tested well - I think this is better
covered in a wider re-factor, rather than here.