Improve Date pages
Use information from the design system to improve the date validation
code.

Currently the date validator erases the user's input if the date doesn't
validate. This is annoying for the user if they had spent time entering
the day and month, only to press an extra key by mistake when entering
the year.

This has been fixed by dropping the dependency on the external date
validation library and implementing the validation in the model.

This increases the complexity of the code but improves the user
experience.