Fix: Submission timestamp issue
Forms can't be submitted without raising an error:
https://sentry.io/organizations/govuk-forms/issues/3548258202/?project=6576261&query=is%3Aunresolved

This happens because `Time.now.zone` was called instead of
`Time.zone.now`. The first returns a string - e.g 'UTC', which then
throws method unknown when the strftime call is made.

This seems a sharp corner of ruby `Time` and rubocop bullied us into
making the error in the first place.