Handle 404 returns
We fixed errors when no data was returned from forms-api but this
didn't solve the issue that if a file was requested and the server
couldn't find the page to serve it would return a 500 error.

This is approach is better, instead of redirecting the user to page not found
and they cant see the url they tried to visit, we simply render the 404 page and
return 404 status code. The user can still the url they clicked/types in
the address bar.