WIP: Change form access to by slug rather than id
This is a WIP branch not intended for release. It pairs with changes to
the API to support accessing forms solely through form_slug instead of
Id.

It's currently not working, throughing an exception:
`ActiveResource::MissingPrefixParam (form_slug prefix_option is missing):`
when accessing a page.

This is probably because of the changes to the Page model prefix in an
attempt to get the runner to call the API correctly.