Remove the page for forms with no start_page
Currently, when a user visits /form/:formid/ they are redirected to
/form/:formid/:form_slug . When a user visits or
/form/:formid/:form_slug they are redirected to the start_page of the
form, or shown a page which says there are no pages for this form.

This commit merges both behaviors. Now a user hitting either route will
be directed to the start_page. If the form does not have a start_page,
they are given a 404.