Add "TEST FORM" to submission email if in preview
We wanted to help organisations identify whether a submitted form came
from a preview or from an actual form. This was a quick fix because we
didn't have any research on what our users needed from previews but we
did now that some people were still accessing the preview form, filling
it and and submitting it to the form submission email inbox. It was
difficult to tell from the email if it was real or not.