Change CYA layout for long answers
When there are long text answers (5000 characyers), the CYO becomes very
long and the page is hard to navigate.

This commit changes the layout from two thirds to full width if any of
the answers contain a long-text question type.

Hopefully this makes the page easier to read and navigate, while still
showing the full text of the answer.