Do not redirect to HTTPS for /ping
Our Puma server is not configured for TLS despite the `config.force_ssl
= true`. Instead if the request has `X-Forwarded-Proto: https` set then
Puma will not redirect to `https` and serve the request. Whenever a
request comes via a level 7 router such as an ALB or the PaaS Router it
sets the header based on the original client protocol.

On AWS the ALB sends a healthcheck to /ping directly and we cannot set
the headers. Stop redirecting /ping to https so we can use it as a
healthcheck for now.

with @alice-carr