Make path /check_your_answers changeable
The route `/check_your_answers` isn't easy to change. For example to
remove the underscores and replace them with hyphens, the route file,
StepFactory and tests would all need to change.

This commit ties these values to a constant,
`StepFactory::CHECK_YOUR_ANSWERS_PAGE_SLUG`. This constant is used to
in the routes, :page_slug constraint, StepFactory functionality and
check_your_answers_controller.