Catch error on loading steps from session
Catch ActiveModel::UnknownAttributeError when loading steps from session,
remove the step from the session storage and return nil.

When form fillers are part way through a journey and form creators
change the question type, the form filler will see an error. The only
way out of this error is to remove the cookie or wait for the session to
be invalidated.

This commit catches the error when loading steps and clears the step
from the storage. The form filler can now continue to fill out the form,
though will have lost the answer to the original question, which is
unavoidable.