Use new live_form route in API
The runner only needs a read-only view of the form currently being
served along with the pages of the form. Currently, two API requests are
made whenever a form page is served - one to fetch the form and one to
fetch the pages.

A new route has been added to the API to return the form including
pages. This commit uses this endpoint to fetch the complete form.