User is redirected to "repeat-submission" if they attempt again
We noticed an issue where users would do the following:

- submit their form answers,
- get a confirmation success page,
- use their browsers "back button" to return to "Check your answers" page
- attempte to submit their answers again

This resulted in submission emails being sent to form processing
teams with no questions or answers.

This commit, moves when a session is cleared until the user lands
on the confirmation page. when it clears the answers for the form the
session changes from

```
{ answers: {
    "4": {
      "1": "This is the answer for the first question of form 4"
    }
  }
}

```

```
{ answers: {
    "4": NULL
  }
}
```