Validate submission email
We need to validate that when a form filler submits their
answer for a form, before attempting to call GOVUK Notify,
we check that the form actually has a submission email.

This should be very rare because live forms should have
a submission email before they are made live. The only real
use case where a submission email maybe blank is when a
form creator is previewing a form and they haven't completed
the add submission email task before they previewed the form. Rather than attempting to call GOVUK Notify without
a submission email and for it to raise an unhandled error we
can instead ignore the call the to the mailer and process
the request as though submission email address was provided
(After all this is just a preview form so it shouldn't really
matter but if form creator has provided the email then we should continue to send the answers to that email address)