WIP - toggling auth headers
Ideally, we dont want to have to stub out auth headers
for every test we create. We should just test the auth headers
in a specific application level unit test.

This commit breaks a lot of specs that use the headers
I suspect we might need to create new application level
test to make sure everything is woring