refactor Context form method
Context.form was extremely confusing because when you looked at current_context.form,
one may assume that it would return the form that the user is filling in. Instead it
would just return the form id. Not only was this confusing but also it meant
that for other form attributes we ended up created attributes in the context to
playback the same attributes we had in the form already. There will be more
refactors but I wanted to try keep them small and targetted.