Refactor the form submission service
It became clear that what we were passing into the
service was different from what we thought was being passed in.

We were passing in the current context to the service but calling it the form when it wasn't actually the Form object.

This refactor helps show the difference of what is passed in and what exactly is being used within the
service