Refactor Step model initialize method
There was a code smell with how Step model was being initialised with so many
params, mostly from other objects but just calling out specific attributes.

This doesn't address all the issues but reduces the number of page attributes that
we need to pass in. As a result it has also meant that we can remove the pages number
method which was used to display the question number.