Add test conditional to form submission email personalisation
We want to vary the content of the form submission email depending on
whether the form was submitted from preview mode or not. We're using the
Notify [optional content][1] feature of their template language.

Unfortunately Notify doesn’t have a way of inverting their conditionals,
so with this method we have to send both two booleans values for both
cases.

We have to be careful that `if true` and `unless true` never have the
same value for the same request, or we will end up with the content for
both preview and live submissions in the same email. As well as
including a comment I've tried to name the values in a way that looks
like a proper conditional, so it's obvious that they are related, but
that also doesn't look like a language construct rather than something
we've added ourselves. `test` matches the language we use in the
template content itself.

Also, stop customising the title based on whether the submission was
from a preview or not, as we are going to do that in the GOV.UK Notify
template instead.

[1]: https://www.notifications.service.gov.uk/using-notify/optional-content