Enable testing GOV.UK Notify mailer
We want to be able to test emails to GOV.UK Notify in an way that
integrates our `FormSubmissionMailer`. The easiest way to do this is to
use the Rails `:test` mailer [[1]], which just puts emails into an Array
at `ActionMailer::Base.deliveries`. Our test config already includes the
usual line to configure the default mailer, but because GovukNotifyRails
sets the default mailer [[2]] itself we need to override the
in a more specific way.

[1]: https://guides.rubyonrails.org/testing.html#testing-your-mailers
[2]: https://github.com/ministryofjustice/govuk_notify_rails/blob/8473dabb2ced1abe040f9aba3481fb5f4addf613/lib/govuk_notify_rails/mailer.rb#L3