Add shell command trace to container run command
It is useful when looking at startup logs to be able to know what command
was run, to understand the context of the log outputs.

This commit enables the `xtrace` option for the shell that runs the
container start command, so that the command to be run is printed to the
logs.