Refactor Log Assertions in Request Specs
Modify the request specs so that they do not make direct assertions on
the final log format. This allows us to modify the logging format without
needing to update these specs each time.

The Controllers either invoke #log_form_event or #log_page_event (or in
the case of the PagesController it goes via the EventLoggingService).
Check that the controllers invoke the necessary higher level function
without making direct assertions on the log format.