Rename helper method question_text_with_optional_suffix
This is ultimately a decorator method for the model but
is doing two lots of things.

Ultimately we will be looking to replace this method
with a combination of Question model method to
return the "Optional" suffix and a component level method
to add the visually hidden span on each question page.