Refactor question_text_with_optional_suffix_inc_mode
Removes logic in the method to decide whether to add suffix to the question text or not.
Instead it uses the models question_text_with_optional_suffix method which will do this
inside the model instead of this helper.

By removing this logic in the helper its made testing a lot simpler because this helper
is only responsible for merging the full question text (whatever that may be) and
adding a suffix for the mode to it.