Component view spec not using correct attributes in test
Rspec was alerting the following warning:

`Checking for expected text of nil is confusing and/or pointless since it will always match. Please specify a string or regexp instead.`

This was result of refactoring we did in https://github.com/alphagov/forms-runner/pull/385 to move the logic of adding
(optional) to the question text to be part of the question model.