Update README.md
Our README.md file in our app repos are (in my opinion) a bit messy and
could use some re-organising to make it the crucial information more
prominent and the rest easier to find.

This commit makes an attempt on the readme for this repo. The changes
include but are not limited to:

- Reorganise sections
- Edit content for clarity and style
- Add more details on running tests and linters
- Add note about normalizing the locales
- Add Sentry section
- Update section on updating Docker files
- Update Notify section
- Update how to contribute section
- Add note about Rakefile
- Remove references to PaaS, we're using AWS for production now
- Fix tests badge
- Fix link to .ruby-version file

Co-authored-by: cmcnallygds <50173207+cmcnallygds@users.noreply.github.com>
Co-authored-by: Tom Isles <tom.iles@engineering.digital.dwp.gov.uk>