conditionally add "X-API-Token" header
We want to conditionally add the header if auth is enabled.
As part of this change, I cleaned up all the tests to no longer
add the token in the stub req headers and replaced them with a
general unit test to check that the base class header works
as we would expect it to during specific conditions.