Add CloudWatch logging for form submissions
This adds an additional logging step to the log event service, which pings Cloudwatch when a form is submitted. As part of the metrics for forms work, we need to keep track of how many times a form gets submitted. We're doing this through Cloudwatch, as it's easy to access from other parts of the app, and requires minimal setup as we're using AWS for other functions as well.