Fix Leaking Email Body When Notify Fails
Two incidents caused the submissions from user forms to be sent to
sentry when the delivery of the submission email failed.

Sentry logs uncaught exceptions. In the submission service, we currently
do not catch any exceptions related to the delivery of the submission
  email. The Notify client throws domain-specific exceptions, which
  provide additional error details.

This pull request resolves the issue by catching the Notify exceptions
and throwing a more generic exception called SubmissionError.
SubmissionError will include the error code from Notify, but no
additional details.

These exceptions will still be sent to sentry since email submission is
critical for our service. We are careful to only handle Notify
exceptions, as we do not want to hide any potential issues by
suppressing exceptions.