Fix broken "edit this question" link
This was caused by a recent refactor where the "Edit question" page url
was changed

I dont think we can write a unit test for this because ultimately we would
need to test if the url actually works and returns the edit page we
expecting to see. I think this would need to be done as part of the
end-to-end tests.

See:
https://github.com/alphagov/forms-admin/pull/659/commits/5a482b5449f6f2bb770bbc2806eca47ce6dff6f6