Add confirmation email form
Adds a basic form to the check your answers page which will pass its
params when the user submits the form. This doesn't include validating
the contents of the form (which will be in a subsequent commit) or doing
anything with the information provided (which will be in a separate PR).