Move submit answers action into CYA controller
When validating the confirmation email later in this PR we'll want to
send users back to the check your answers page if the email is invalid.
This means we'll have to render the local/instance variables used by the
check your answers view in both actions.
Combining the check your answers and submit answers controllers will let
us do this without duplicating code between the controllers.