Create email confirmation form model
Creates a form model EmailConfirmationForm which will be where we handle
the validations for the email confirmation form. Adds translations to
the form fields using the helpers in the govuk_design_system_formbuilder
gem.

Because we're using the govuk_design_system_formbuilder helpers, we need
to add a line to the i18n-tasks config to prevent it falsely labelling
these translation keys as unused.

Includes a change suggested by @aliuk2012 to update the field values to
avoid confusion between booleans and strings pretending to be boolean.

Co-authored-by: Alistair Laing <aliuk2012@users.noreply.github.com>