do not attempt to send confirmation email if form what happens next is blank
We need the forms what happens next as we use it to form most of the email body.
Without this theres little point of sending the confirmation email. This
should only ever happen when a form is a draft and a form creator is attempting
to see what confirmation email would look like.