do not attempt to send confirmation email if form has no contact details
We need the forms support contact details as we use it in the email body.
Without this theres little point of sending the confirmation email.
This should only ever happen when a form is a draft and a form creator is
attempting to see what confirmation email would look like.