Move logging configuration to its own file
Move the logging configuration from application.rb to an initializer.
This makes it easier to apply it to test and production but not
development environment.

The desired effect is that when developing locally, the default Rails
logger is used. Our default logger, lograge, outputs structured log
messages that are optimised for machine readability, rather than human
readability.

If a developer wants to see what the production logging looks like they
can change the RAILS_ENV environment variable or comment out the
conditional in the initializer.

This also means we can be aware of what we are missing in our lograge
configuration that we could be getting from Rails.