Add notification reference for confirmation email
We want to be able to find the confirmation email for a form filler in
the GOV.UK Notify API. This is currently for our end to end tests, so we
can check that a notification is actually queued.

This commit adds a notification reference for the confirmation email,
similar to how we already have a notification reference for the email
with the submission. To make it easier for any humans looking at the
Notify logs to associate the two emails together, we give them both the
same random prefix, and differentiate with a human readable suffix.