enforce 4-digit year field for date answer types
Year fields should not accept two digits. Instead of checking the length of year which would mean that 0000 to 0999 would be valid
dates, I instead check that year is between 1000 and 9999. This isn't exactly ideal. Having checked this with a designer, we decided to do this check and allow us to later add more refined validations which could be configured by form creators.

This follows the GOV.UK Design System pattern to ensure years are 4-digit numbers
https://design-system.service.gov.uk/components/date-input/#if-the-date-is-incomplete