Add fill and submit form feature test
Adds a new feature spec which covers the basic form filler journey. It only contains one question, which is answered without triggering any validation, and the form is submitted without opting for an email confirmation.

There's also a tweak to our use of the Axe clean accessibility checks, which was flagging up our use of `aria-expanded="false"`. It looks like this should be passing, and the Axe rule hasn't been changed yet to reflect this. As a result, we've ignored any triggering of this rule on radio buttons. See https://github.com/alphagov/govuk-frontend/issues/979 for more details