Add session_id_hash to domain logging
Logs for domain events don't currently include session_id_hash.

This commit adds the hash to the logged JSON for these events.

To keep the hashes consistent, it centralises the code used to hash the
session id.

The hash is not stored anywhere so a request with a log event will hash
the session more than once.

The digest used to hash the session_id is quick. If performance became
an issue, we could use middleware to hash the session_id once and store
it in the request object itself.

Adding a new class, SessionHashser, probably isn't the solution to
standardising the request logging but keeping the scope of this PR small
will make refactoring logging easier in the future.

Fix rubocop issue in spec