Fix visible environment tag
We updated the tags to reflect the new tag casing rules in version 5 of
govuk-frontend. Unfortunately the 'production' text was not updated to
'Production' everywhere, so the logic for hiding the environment tag in
production broke.

This commit fixes the problem, and updates the 'production' check to
look at the translation directly, so that the environment check will
stay up to date with any future changes to this text.