Remove line installing Bundler
We don't need to explicityly install a specific version of Bundler in
our Dockerfile, because the `ruby` base image includes Bundler already,
and running `bundler install` will install the correct version if a
different version to the latest was used to generate the `Gemfile.lock`
file.

As we don't need the line, and having it makes bumping the version of
Bundler used more complicated, this commit removes it.