replace custom health check with rails default
rails 7.1 introduced a health check endpoint by default. This default endpoint is called `/up`. Instead of returning 200 with the body response of `PONG`, it will return 200 if the app is healthy and it will output a html page with the background color set to green. If there is an issue it will return 500 error and the html page will have a red background.

We will need to update our infrastructure to use /up instead `/ping` and then we can remove our custom heartbeat controller.

A simple test was added as part of this commit, not to test Rails implementation but as a means to check the functionality is present in the application and should fail if someone deletes the route in error.

see https://github.com/rails/rails/blob/f481353150c100dbd944d5aa84f2480264cd723c/railties/lib/rails/health_controller.rb