Refactor test of submission error
I found that this test was giving a false positive, because an exception
was being raised earlier than expected, because some prerequisistes for
the submit_answers action were missing.

This commit makes the following changes:

- Use factory instead of hard-coding form data
- Add pages to live form response data
- Provide confirmation_email_form params in request
- Build up session context properly
- Ensure form_submission_service mock is constructed fully

The test still passes, but now I'm more confident it's testing the thing
we want it to; if the test is changed to not raise an exception, the test
fails (when previously it would pass).