Add helper for subscribing to action controller notifications
We're using the same bit of code to subscribe to `process_action`
notifications and capture the payload in a couple of different tests
now, this commit creates a shared helper to reduce duplication.

I wanted the subscriber to be global, so that we don't need to
subscribe/unsubscribe for each example (apparently there is a
performance penalty when doing so [[1]]. Having the subscriber as some
kind of class variable seemed the obvious way to go to do that.

Note that this uses class variables, which are not recommended by the
Ruby FAQ/style guide [[2], [3]]. However, using class instance variables
doesn't seem to work with RSpec (I get errors about the variables not
being available in a example context, only within a specific example,
which doesn't make much sense to me), so we just ignore the cop.

[1]: https://api.rubyonrails.org/classes/ActiveSupport/Notifications.html#module-ActiveSupport::Notifications-label-Temporary+Subscriptions
[2]: https://www.ruby-lang.org/en/documentation/faq/8/
[3]: https://rubystyle.guide/#no-class-vars