Log Question_Number
The `question_number` field was omitted during the recent refactor to
logging[[1]]. This adds it back in. Note that `question_number` is the
sequential position of the question with reference to the form (e.g. the
first or second question in the form) and the `page_id` is the unique
identifier of the page in our records.

[1]: https://github.com/alphagov/forms-runner/pull/598