Add log_rescued_exception helper
We want to log some exceptions even if we handle them in a controller
action.

Unfortunately I couldn't find an easy way to hook into the standard
Rails exception logging mechanisms at the same time as using `rescue` or
`rescue_from`; so this commit adds a new method to our application
controller that has the same behaviour.