Prevent Duplicate Event Logs
`tap` mutates the hash and then returns it. This meant that the
additional event data was being added to the logging_context and
logged by LogRage as well as the additional log line created by the
EventLogger[[1]]. This is a quick fix to prevent the additional event
context being logged twice which is causing problems for Anne's
analysis.

I think we should explore removing the EventLogger and just adding the
event data to the logging_context and let LogRage log it; this however
requires more refactoring and I'd like to fix this for Anne before it
causes too much bother.

[1]
https://github.com/alphagov/forms-runner/blob/main/app/lib/event_logger.rb#L3