Temporarily add notification-id for end to end tests
Our end to end tests look for the hidden field `#notification-id` to
know how to check whether delivery of the form submission email was
successful (see [[1]]). However, in the previous commit we rename that
field `#submission-email-reference`.

To allow the end to end tests to continue working while the changes in
the previous commit are deployed, this commit temporarily restores the
original field alongside the new one, using the 'expand and contract'
pattern. Once the changes in the previous commit have been deployed to
all environments and the end to end tests updated (see [[2]]), we can
revert this commit.

[1]: https://github.com/alphagov/forms-e2e-tests/blob/9770d3dff373cf97dd4cc94f3fdb81748d220e1d/spec/support/feature_helpers.rb#L160
[2]: https://github.com/alphagov/forms-e2e-tests/pull/41