Move email_sent into session
In the previous commit, we used the session to persist the submission
reference between the submit_answers action and the `submitted` action.

Since we now have a way to persist data between these actions, we can
reuse it to pass the `email_sent` boolean to the submitted view.

Previously we used a flash message for this, which isn't really the
intended use case for flash messages, and also meant that refreshing the
confirmation page made the `email_sent`-dependent content disappear from
the page. This will no longer happen.