Clear submission details on first page save
We store the submission details in a part of the session that isn't
cleared when the form submits. This is necessary for the submission
details to be shown on the confirmation page.

It could get confusing to keep this data in the session, since if the
user starts filling in the form again their session will include
information about multiple different form submissions.

This change clears the submission details from the session when the user
starts the form again, so that the user only ever has information about
a single submission per form in their session.

It uses the first save as our definition of 'starting the form', since
this is the point where data for the new form filling attempt gets
saved to the session and the discrepancy would otherwise arise. This
also aligns with the point where we log form starts in our metrics.