Add feature flag for submission references
This adds a flag in the settings file to turn on or off the reference
numbers. Note that this will stop the reference numbers being displayed
or logged, but will not stop them being stored and cleared from the
session.

This is necessary so that we can do the email work (trello #1361) as a
separate PR. Without this, some users would be shown a reference number
that neither they nor the form processor would be emailed, so it would
be useless.