Add reference storage methods to context
This commit adds methods for storing and retrieving submission
references from the session.

Since these references will need to be shown after the user has
submitted their form, the reference needs to be stored in a part of the
session which is not cleared by the `clear` method which runs after
every submission.