Generate and store reference in controller
Adds a method for generating reference numbers to the
FormSubmissionService, then returns this in the `submit` method. The
Check Your Answers controller then gets the returned reference and
stores it in the session.

Uses SecureRandom.base38 to generate the string, since this reduces the
risk that the generated ID will have ambiguous characters (O/0 etc) in
it, and makes it uppercase. These choices should make it easy for a form
filler to do things like write their reference down on paper and quote
it over the phone.