Add reference numbers to submission email
This adds the submission reference number to the email sent to
processors when the form is submitted. It also includes a boolean
personalisation field `include_submission_reference` which determines
whether the reference number and its associated text should be shown in
the email. This allows us to keep the email template change behind the
`reference_numbers_enabled` feature flag.