Update en.yml to make confirmation email hint more specific
Following user feedback, this changes the hint about how we use the email address you provide for a confirmation email to make it ever so slightly more specific. This is to help avoid people thinking that their email address won't be used for something else if it's been collected earlier in the form.

We need to think about this a bit more to improve it further but this should help a bit for now.