Remove redundant main component
In the previous commit, we removed the watermark from the main
component. Removing the watermark from this component meant that the
only logic left in the component was for removing the id if the
component was being previewed locally - it had no user-facing function
at all.

This commit removes the component altogether and replaces it with a
simple `main` tag.