Allow BR tags in markdown
We were deliberately restrictive when originally implementing our
markdown scrubber, so that we could be sure only the necessary HTML tags
could be rendered on the site.

Form creators have valid reasons for using <br> tags in their markdown -
for formatting addresses in particular - so this commit adds <br> to the
scrubber's allow list.