Show error page when user tries to submit incomplete form
It is possible for a user to submit an incomplete form after they have
already submitted a form if they duplicate the tab with the
"Check your anwsers" page before submitting. If they then continue to
answer at least one question in a new submission, they were able to
go back to the already open "Check your answers" page and submit the
incomplete form.

This commit causes an error page to be shown if the user manages to
click the submit button on the check your answers page, but all
questions have not been answered. The error page has a link to go back
to where they should be in the form.

There is a small chance that a user could see this error page if they
haven't already submitted a form in the case that a new version
of the form is made live with additional questions at the end when a
user is on the check your answers page. This is because we check
whether the last question they answered is the last question in the
form before we allow it to be submitted.

Sending an event to Sentry when a user attempts to submit an
incomplete form has been removed as this was added as a temporary
measure to be informed if this happened again before we fixed the bug,
but we continue to log an event to Splunk.